Refactor imports in deep_flow modules to align with new structure
