+++ pipelex/deep_flow/test_extras/wf_test_child_pipe.py
@@ -3,12 +3,12 @@ from temporalio import workflow
 with workflow.unsafe.imports_passed_through():
     from pipelex.cogt.llm.llm_setting import LLMSettingChoices
     from pipelex.core.pipes.pipe_output import PipeOutput  # noqa: TC001
-    from pipelex.core.pipes.pipe_run_params import PipeRunParamsFactory
     from pipelex.deep_flow.log_temporal import workflow_log
     from pipelex.deep_flow.test_extras.models_for_tests import Mock
     from pipelex.hub import get_pipe_router
-    from pipelex.pipe_operators.pipe_llm import PipeLLM
-    from pipelex.pipe_operators.pipe_llm_prompt import PipeLLMPrompt
+    from pipelex.pipe_operators.llm.pipe_llm import PipeLLM
+    from pipelex.pipe_operators.llm.pipe_llm_prompt import PipeLLMPrompt
+    from pipelex.pipe_run.pipe_run_params_factory import PipeRunParamsFactory
     from pipelex.pipe_works.pipe_job_factory import PipeJobFactory
     from pipelex.pipeline.job_metadata import JobMetadata
